chore: remove cynic-introspection dependency from engine (#718)
The grafbase engine doesn't really need cynic-introspection (or cynic)
for anything. It's only a dependency because `parser-graphql` wants to
make a Registry out of some cynic-introspection structs, and the orphan
rule prevented the From impls from being in `parser-graphql`.

Cynic isn't the heaviest dep but it's not trivial, so I decided to put
the time into extracting it.