chore: split ContextBase up  (#730)
I'm working on GB-4762. This issue is caused by the ResolverChainNode
not being set up correctly in deferred contexts. It's quite hard to do
this correctly in a deferred context because it contains references to
its parent, and without running a bunch of execution calls again it's
hard to get the lifetimes to line up.

It is also a bit of a mess of a struct - it's got a bunch of Options
that are always set depending on the context (which isn't documented so
you just sort of have to know) and it's got a lot of overlap with
`ContextBase`. So I'm probably going to remove it and merge the bits we
need into `ContextBase`.

First though, I need to split `ContextBase` in two: it's used both when
we're resolving fields and when we're resolving selection sets, and the
data we have available is different in both those contexts. And that's
what this PR does.

I've also:

- Moved the response & deferred workloads into QueryEnv because they are
more or less global to the query.
- Removed some custom directive & directive derive stuff. We weren't
using it and it was a pain to update for this change.
- I've added a `ContextExt` trait where functions shared between both
context types can live.
- Removed most of the ResolverData stuff that has been deprecated for
most of the year. I haven't removed from the Resolver type just in case
we still have schemas that are using it - but they will now error if
they hit the path that needs it.

My next PR will merge the remaining ResolverChainNode stuff in to these
structs.