chore: merge ResolverChainNode into Context (#731)
I'm working on GB-4762. This issue is caused by the ResolverChainNode
not being set up correctly in deferred contexts. It's quite hard to do
this correctly in a deferred context because it contains references to
its parent, and without running a bunch of execution calls again it's
hard to get the lifetimes to line up.

It is also a bit of a mess of a struct - it's got a bunch of Options
that are always set depending on the context (which isn't documented so
you just sort of have to know) and it's got a lot of overlap with
`ContextBase`. So I've removed it and merged the relevant bits into the
respective Context structs.

This results in a lot less Options, and some more specific types. Other
changes I needed to make:

- Got rid of the `Context` alias for `ContextField` so I could create a
`Context` trait.
- Added ContextList for resolving lists (it was using
ContextSelectionSet before, but there might not actually be a selection
set)
- Made a `ContextSelectionSetLegacy` for the legacy async_graphql
resolution stuff, because `ContextSelectionSet` didn't _quite_ work for
it.
- Probably some other stuff I've forgotten to mention.