fix(connector): Allow emitting input field for nullable composite key (#761)
This fixes an edge case that might seem small, but would lead to
catastrophic results in the worst case. Let's imagine a table:

```sql
CREATE TABLE "user" (
  name VARCHAR(255) NOT NULL,
  email VARCHAR(255) NULL,
  CONSTRAINT foo_key UNIQUE (name, email)
);
```

The input type would be something like this:

```graphql
query {
  user(by: { nameEmail: { name: "Naukio", email: null } }) {
    name
    email
  }
}
```

This would trigger a query with `WHERE name = 'Naukio' AND email IS
NULL`. Now, the column being nullable, the generated graphql type allows
one to emit the value completely, so this is also allowed:

```graphql
query {
  user(by: { nameEmail: { name: "Naukio" } }) {
    name
    email
  }
}
```

Before the fix, it renders a query `WHERE name = 'Naukio'`, which then
could return more than one rows, leading to undefined behavior. And
what's worse, when we implement the deletion or updates of one item,
this would delete/update more than one item in the database.

The fix maps all fields _not defined_ in the query that are part of the
constraint, and adds `AND field IS NULL` for the fields missing in the
input.

Closes: GB-4909