fix(cli): implement grafbase binary selection on bun
When you try to run the grafbase CLI from npm with bun, for example with
`bunx grafbase`, you get errors that look like this (linux):

```
error: Failed to run "grafbase" due to error InvalidExe
```

or like this (MacOS):

```
grafbase dev
/Users/fredrik/.bun/bin/grafbase: line 1: This: command not found
```

The `grafbase` npm package relies on a postinstall npm lifecycle hook to
link or copy the right platform-native binary to
node_modules/grafbase/grafbase. Where bun is different is that it does
not run postinstall hooks, or any lifecycle scripts for that matter
(source: https://bun.sh/docs/install/lifecycle).

This commit is a slightly acrobatic attempt at solving this problem by
replacing the placeholder `grafbase` file in the `grafbase` npm package
with a script that will overwrite itself with the correct binary on
first invocation. The script simply runs the postinstall hook as a
regular import, since bun will not run it for us.

Part of GB-4936