examples: add a KV + typed resolver and authorizer example (#807)
I haven't found existing examples (in `/examples`) using KV, so this
seemed like an opportunity. The original purpose of this PR is to have a
working example of typed resolvers and authorizers with the recently
added types from SDK, so we can link to it in the changelog.

It is very very basic and not much of a functioning app, feedback
welcome.