feat(cli): implement typescript code generation for resolvers
First iteration of the implementation of the GraphQL SDL -> TypeScript
resolver types part of Typed Resolvers.

This is a new standalone crate in `cli`. It does not reuse existing
infrastructure because the amount of analysis required on the schema is
very limited: we do not want to do extra validation because we want to
tolerate fairly broken schemas and still generate code,and code
generation has to be fast for integration in the `gb dev` loop.

Unlike graphql-code-generator, we do not generate types for the resolver
functions. By design, the generated module has one TypeScript type per
GraphQL type. They have the same name (except TS reserved keywords), and
they are in the same order in the generated file as in the source file.

We do not want types for the resolver functions because we are going to
try using the resolvers as the source of truth (RFC section pending).
Resolver function types can easily be added later if that approach does
not pan out.

Not implemented yet:

- Custom scalar types. They are typed as `any`. We probably want to have
  a configurable location for the generated code to import them from.
- More subtle interface typing. The current approach is very naive, it
  only includes the interface fields.

closes GB-5146