feat(sdk): add new ResolverFn export
This is a convenience type for resolver functions. It is generic because
resolvers have different parent, argument and return types. It works for
both sync and async resolvers. This will be used by generated code as part of the typed
resolvers work, so it is not meant primarily for direct end-user consumption.

Related to https://github.com/grafbase/grafbase/pull/843 but not
blocking.

first part of GB-5163