fix: silence federation introspection queries in dev command (#869)
If you're running rover in development it'll poll your subgraph every 2
seconds. If that subgraph is `grafbase dev` then it's quite noisy: a
constant stream of `query SubgraphIntrospectQuery 0ns` messages in your
console.

This PR updates the `is_operation_introspection` query to consider the
`_service` field as part of introspection, which should silence those
queries like we do with normal introspection.