feat: @join directive (#862)
I've been working on a test case for the OpenAPI<->Subgraph integration
where I've got a Locations API & a Reviews API. With the integration as
it currently stands I've managed to add a `location` field into the
`Review` entity quite easily. What's currently quite difficult is
extending the `Location` type with endpoints from the Reviews API.

My reviews API has an endpoint `location_reviews/:id` which I'd like to
add to the `Location` type as `reviewsForLocation: [Review]`. The
OpenAPI connector exposes this endpoint as a field on the `Query` type:
`locationReviews(locationId: String)`. But there's no way to add that
`Query` field onto the `Location` type.

I've solved that by adding the `@join` directive that we've spoke about
in the past, as it does basically what we want. So for my test case I
can add the following to the reviews subgraph schema:

```graphql
type Location @key(fields: "id") {
  id: String!
  reviewsForLocation: [Review] @join(select: "locationReviews(locationId: $id)")
}
```

The syntax is a GraphQL field selection, and all the fields of the
current type are implicitly available as variables. It'll automatically
infer an `@requires` directive from the used variables, so users
shouldn't have to worry about making sure they query for those fields.
(`@requires` doesn't work for all our connectors at the moment, but we
can fix that when/if it becomes a problem).

This is very much a first pass - you can't pass nested fields in to the
sub-query, there's no way to pass arguments from the field to the
sub-query and you can only query top level fields - no nesting at the
moment. We can handle those later if needed though - I have ideas for
all of them so pretty sure they're all possible.

I've skipped validating the query much in this PR as it's already quite
big. Will do that & adding to the SDK in separate PRs.