feat: extend openapi types with federation keys (#867)
The OpenAPI connector attempts to infer federation keys from the shape
of the underlying API. If an endpoint returns a named type from the
schema, and has a URL parameter that clearly maps to some field on that
type, then we can easily generate a resolver. This works OK for some
schemas but there are a lot of edge cases where it just doesn't.

For those cases, we want users to be able to specify how a key should be
resolved when the `_entities` field is called with that key. This PR
does that by adding a `select` field to the `@key` directive - which
takes the same input as the `@join` directive and is hooked up to the
same resolution mechanism.