composition: return structured federated graphs
Until this commit, the output of composition is a GraphQL SDL string
without any federation-specific metadata attached.

This commit implements a much more structured output format that is
designed to be serialized. The format is defined in a new crate:
`grafbase-federated-graph`. It is a normalized representation of the
schema, with all strings and all field types deduplicated.

The composition crate is reorganized for clarity and to make that
structured output possible. See the comments in the diff for more
details.

The new `grafbase-federated-graph` crate also defines a function to
render a structured federated graph to a GraphQL SDL string.