ci: try collector job
Problem:

    If we make the darwin and linux jobs required, then CI stays yellow when you make changes inside cli/.
    If we make the jobs inside the darwin and linux matrices required, then CI stays yellow when you make changes outside cli/.

It seems that mixing required checks and matrices just doesn't work, there are github discussions, issues and articles about this.

Solution: the "collector job" pattern. We already have a job that depends on all the build jobs succeeding: the one posting the PR comment with the built binaries. The solution is to make this job required, make it always run, and artificially fail if any of the build jobs failed or was cancelled. That way we get the expected behaviour of the build job's success being required when the CLI actions run, with CLI jobs skipping still working when no changes have been made inside cli/.