feat: GraphQL schema validation checks #933
Based on the 2021 version of the GraphQL spec. This should cover the
spec pretty much exhaustively. The error messages are using formulations
users would be familiar with.

The diff is big but a large part is due to tests. For review, I would
suggest reading following the control flow starting from lib.rs. It is
very linear.