feat: federation `@external` directive (#901)
This adds support for the federation `@external` directive to the
engine. This indicates that a field or object is external to the current
sub-graph. There's no actual logic required in the subgraph for this, we
just need to expose it as part of the schema so the federated graph can
do its thing.