Add engine-v2 graph
Adding a base structure to work on, will definitely iterate a lot on it.
It's almost a copy paste from the federated-graph for now.

I don't think unit tests for the conversion itself are meaningful. This should
be tested through integration tests like composition. But instead of
checking a SDL output we would check introspection and that the engine
behaves as expected.