fix: don't assume templates are SDL based in init (#915)
`grafbase init` can set up a project from a template. The output assumed
that all the templates were written using SDL though, and always pointed
the user at a `schema.graphql` file. This makes it a little smarter:
it'll now do the right thing depending on the template.

Unfortunately had to get rid of a `tokio::main` attr because it didn't
like return values, but 🤷‍♂️.