chore: make `MockGraphQlServer` generic over any schema
For testing federation we're going to need a variety of different GraphQL
servers in our integration tests.  This commit is prep work for that: it makes
the MockGraphQlServer able to support any `async_graphql` schema.

A follow up PR will add some that we can use for federation.

Part of GB-5337