composition: release to crates.io (#946)
This is the bare minimum change to get the crates to crates.io. We
should do this now in the spirit of release early, release often, and
improve from there. That will already enable depending on the crates.io
version of the crates instead of the git repo in other places.