feat(composition): implement entity interfaces
The implementation is pretty long, I would suggest getting familiar with
the docs and reading tests first to review.

This commit contains a change in control flow in how we compose fields
of objects. At a high level, the change here is the following:

- Before, we were iterating every object type in all subgraphs not
  thinking about their fields, then iterating all object fields in a
  second pass.
- Now, for every object, we iterate the fields for that object when
  composing the object in the first pass.

The reason for the previous behaviour was that we were expecting the second pass to need all the results from the first pass (e.g. to resolve field types), but it turns out we do well without, we don't really accumulate data during composition and composition IR is read only, which is a good property.

The reason for this change now is that for entity interfaces, we really want to compose the fields in a context where we know that we are dealing with fields that are part of an entity interface, so in this PR, we stop throwing away that context.

closes GB-5229