fix: compile errors in integration tests (#1052)
We merged the `engine` & `cli` workspaces, so integration-tests now gets
compiled in IDEs with the `sqlite` flag enabled in `dynamodb`. This
results in compile errors, which is a bit annoying.

This PR works around for now by just making the signatures the same when
feature is enabled & disabled. The tests won't work with the flag
enabled, but better than not compiling.