fix: dont wait for poll timeout if cli dies
Last month we had a bunch of long CLI CI runs.  In one case this was because
the CLI failed to start up in every single test, which is not a case we detect
- we wait for a 30 second timeout before ending the test.  Since we retry each
test 30 times, this is a lot of waiting around.

This commit updates the `poll_endpoint` function to watch for all the current
commands dying - if that happens there's no point in continuing to poll.

Have tested this with some simulated failures locally, and it does end the
tests significantly faster.


the CLI tests all being retried 3 times