fix: cancel in progress CLI builds (#1087)
We currently don't have cancel-in-progress turned on for our CLI CI
builds, which means if you push a new commit you have to wait for any
existing CI runs to finish before your latest commit will get built.

Not sure if there's a good reason for this, but I can't think of one, so
maybe lets try enabling `cancel-in-progress`? Shout if you know of a
reason why this will be terrible.