cli: release 0.48.0 (#1092)
Features

- A new `grafbase check` command to run schema checks in federated and
single graph projects.
- `grafbase publish` can now take the schema SDL to publish from stdin
as an alternative to a file with the `--schema` parameter. This enables
piping the output of introspect into publish. (#1083)
- In all commands that take a `PROJECT_REF` argument
(account/project@branch), the branch is now optional. It will default to
the main branch of the project. (#1081)
- `grafbase init` now generates an initial configuration that takes the
latest changes in @@grafbase/sdk into account.

Fixes

- File watcher fixes (#1064). `grafbase dev`, was watching too many
files, leading to unnecessary reloads.