cli: release 0.48.1 (#1097)
Features

- The `grafbase introspect` command can now take a `--dev` flag to
introspect the local project's schema. It is mutually exclusive with the
`--url` flag. (#1095)

Fixes

- Due to the recent added support for configuration at the repository
root, the generated TS code for typed resolvers was generated at the
repository root, even when the configuration lived in the ./grafbase
directory. This releases fixes that behaviour: the `generated` directory
will be at the same level as the configuration. (#1093)