feat: add headers to federated-graph
As we discovered in the demo this morning, we need a way to specify headers
that should be sent to a subgraph in `engine-v2`.	

This PR adds fields for storing this information into `federated-graph`.  I'm
not using them yet, wanted to get this out for review to make sure we're all
agreed that this is the right place to put them.

Part of GB-5590