chore: split config out of servers.rs in cli (#1136)
I'm trying to update the watcher code in the CLI to better support the
federated dev command.

This is a bit of a preparatory tidy for that work: splitting
configuration stuff out of `servers.rs` and into its own module,
renaming some functions etc. Not essential, but I'm hoping it makes my
job a tiny bit easier

Should be no functional changes in here.