chore: optimize build time (#1147)
I've been running a lot of CLI release builds on my machine recently and
it is incredibly slow to compile - about 4 minutes, even if I've only
changed a couple of the CLI crates and all the dependencies are already
built. CI is also pretty slow - 30ish minutes for a full run, much of
which is spent compiling for the various platforms.

This PR speeds both those things. CI runs are down to ~18 minutes,
release builds on my machine now take less than 2 minutes.

This does increase the size of the built CLI a bit - up from 38MB to
60-70MB. I kinda think that's fine, but comment if you don't?