composition: implement non-default root type support
Concretely, we now take into account the schema definitions (e.g.
`schema { query: MyQuery }`).

There are subtleties in the rules:

- If there is a custom mutation root (or query, or subscription) and a
  type that has the default name for the root (Query, Mutation or
  Subscription), it is a hard error, to align with expectations.
- If a type is a root in some subgraphs but not in others, it will be
  handled as a regular type in the subgraphs where it isn't the root and
  as the corresponding root where it is. We are better behaved than
  other implementations in that regard.

There are many changes in this PR that look unrelated to the matter at
hand. This is because until now, we could assume that all the types with
the same name compose together (and their fields too) across subgraphs.
This is no longer the case, take these two subgraphs for example:

```graphql

schema {
  query: CocoQuery
}

type CocoQuery {
  coconutYoghurtVariants: [String]
}

type Query {
  soyYoghurtsVariants: [String]
}

```

The `CocoQuery` and `Query` type will compose together as the `Query`
type of the federated schema. This broke a number of assumptions about
composition, so it had to be addressed. Among other things, special
treatment for root types reorders them in the federated schema, hence
the snapshot changes.

closes GB-5673