composition: reinstate broken assumption about interned strings
A while ago, we introduced a separate string interning map for static
strings that may not show up in the subgraphs but are part of the
federated graph and as such, need to be in `FederatedGraph.strings`.

The problem is that the same string can end up in both the map for
strings from the subgraphs, and, if the static version was inserted
before, in the second map for static strings.

The important consequence of this is that you could end up with two
`StringId`s that are not equal representing the same string, which
defeats the point of one big reason for interning in the first place:
cheap and reliable comparisons.

This commit fixes that by not trying to be smart about converting
`subgraphs::StringId`s into `federated::StringId`s and rehashing the
string themselves instead.