parser-sdl: fix extending type causing duplicate type definition error (#1143)
Roughly, there are three scenarios for extend type:

- extend type Query / extend type Mutation -> these are special cases,
there are tests
- extend types for types from connectors -> also tested
- extend type for types defined in the config / sdl -> that's where the
issue applies

In that last point, two subcases: extending a model and extending a type
that is returned by a resolver extending a connector type. The first use
case is being sunset. The second use case is tested and fixed in this
PR.

closes GB-5067

# Type of change

- [ ] 💔 Breaking
- [ ] 🚀 Feature
- [x] 🐛 Fix
- [ ] 🛠️ Tooling
- [ ] 🔨 Refactoring
- [ ] 🧪 Test
- [ ] 📦 Dependency
- [ ] 📖 Requires documentation update