fix: swap ts-node for tsx.
A user found another problem with ts-nodes ability to work with CJS & ESM
files, which appears to be [this
issue](https://github.com/TypeStrong/ts-node/issues/1997).  Swaping to tsx
instead seems to fix this and removes the need for having two different copies
of the parse-config script.

Depends on https://github.com/grafbase/api/pull/3073 being pulled in.