fix(cli): generate TS for fields w/ grafbase scalar type
The SDL generated from user configuration does not contain custom scalar
definitions for builtin types like DateTime and IPAddress, so the code
generation logic could not resolve the types of these fields and left
them out of the generated code.

The short term fix in this PR is to integrate those in the codegen
worker's understanding of builtin scalars.

The long term fix will come with basing TS code generation on the whole
registry, which we want to do anyway for a host of other reasons.

closes GB-5694