operation-checks: initial crate with breaking up of operations into fields (#1191)
This PR adds a crate in engine/crates. The crate is called
`operation-checks` and is meant to hold the business logic for operation
checks.

At a high level, within a schema check, the operation check step does
the following work:

1. Gathers the fields that are actively used by API consumers, with
thresholds based on user settings. This can be done in two ways:

a. Taking already aggregated field data directly from an analytics
database.

b. Taking aggregated operation-level data, and splitting it into
field-level data.

2. Diff the old and the new API schema. The diffing part is already
implemented in the graphql-schema-diff crate.

3. Determine if any of the breaking changes in the diff would apply to
fields actively used by clients (determined in (1.)).

This PR implements bullet point 1.b.

Two tradeoffs lead to this approach:

- No splitting of field-level data for now because it turned out more
challenging than expected, and we are already aggregating at the
operation level, so the fine grained approach may not turn out to be
necessary. The approach in this PR is also significantly less work to
get a first version of the feature out.
- Not depending on engine-v2 for the splitting, but on GraphQL ASTs
instead: the API is less of a moving target and I did not want to burden
engine-v2 with another consumer with its own needs. For example, we want
to error if a query selects a field not present in the schema in engine
v2, but it is a completely expected situation in operation checks, since
we work with queries that were sent to past versions of the API.

Next PR in this crate will implement part 3 as a function taking the
diff and the usage counts from this PR as arguments. It will output
operation check warnings and errors.

closes GB-5692